Add release_disabled field app- and unit- tests
Previously if suite.ini has a release_disabled field it works only for
tarantool tests. This commit allows to use it for app- and unittests
suites.

Despite unittest does not take tarantool binary, unittests get version
(release/debug) from the tarantool binary information.

For instance,

Add line to test-unit/suite.ini:
`release_disabled = broken_unicode.test`

On debug version:
`[001] test-unit/broken_unicode.test                                   [ pass ]`

On release version:
`[001] test-unit/broken_unicode.test                                   [ disabled ]`

Closes: #199
